% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db2_connect.R
\name{db2_connect}
\alias{db2_connect}
\title{Connect to DB2}
\usage{
db2_connect(db_name = NULL, username = NULL, password = NULL,
  server_address = NULL, driver_path = NULL)
}
\arguments{
\item{db_name}{name of database}

\item{username}{name of user}

\item{password}{user password}

\item{server_address}{address to server}

\item{driver_path}{Path to JDBC DB2 driver}
}
\description{
Set up connection to DB2 server
}
