% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db2_connect.R
\name{db2_list_columns}
\alias{db2_list_columns}
\title{List columns in DB2 table}
\usage{
db2_list_columns(conn, schema, table)
}
\arguments{
\item{conn}{database connection object}

\item{schema}{name of schema}

\item{table}{name of table}
}
\description{
Lists columns and colums types in a DB2 table
}
