% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db2_connect.R
\name{db2_list_objects}
\alias{db2_list_objects}
\title{List objects in DB2 database}
\usage{
db2_list_objects(conn, schema = NULL)
}
\arguments{
\item{conn}{database connection object}

\item{schema}{name of schema, specified to list tables in a schema. 
Leave as NULL to show list of shemas in database.}
}
\description{
Returns list of objects (schemas or tables) in a DB2 database.
}
