% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db2_connect.R
\name{db2_query}
\alias{db2_query}
\title{Get Query}
\usage{
db2_query(conn, qu)
}
\arguments{
\item{conn}{db2 connection}

\item{qu}{character string with SQL query}
}
\description{
Get data from DB2 connection using SQL query
}
